better code layout and new header fn
