+++ kitsune/src/http/openapi.rs
@@ -91,8 +91,8 @@ struct CommonApiDocs;
         mastodon::api::v1::accounts::verify_credentials::get,
         mastodon::api::v1::apps::post,
         mastodon::api::v1::follow_requests::get,
-        mastodon::api::v1::follow_requests::accept,
-        mastodon::api::v1::follow_requests::reject,
+        mastodon::api::v1::follow_requests::accept::post,
+        mastodon::api::v1::follow_requests::reject::post,
         mastodon::api::v1::instance::get,
         mastodon::api::v1::media::post,
         mastodon::api::v1::media::put,
