+++ kitsune/src/http/pagination.rs
@@ -26,10 +26,46 @@ where
             .collect::<Vec<String>>()
             .join(", ");
 
-        // as long as we pass valid links this should never panic
-        res.headers_mut()
-            .insert("Link", HeaderValue::from_str(&value).unwrap());
+        // TODO: log an error when HeaderValue conversion fails
+        if let Ok(header_value) = HeaderValue::from_str(&value) {
+            res.headers_mut().insert("Link", header_value);
+        }
 
         Ok(res)
+    }
+}
+
+pub fn new_link_header<I, D: Display, F: Fn(&I) -> D>(
+    collection: &Vec<I>,
+    limit: usize,
+    base_url: &str,
+    get_key: F,
+) -> Option<Link<Vec<(&'static str, String)>>> {
+    if collection.is_empty() {
+        None
+    } else {
+        let next = (
+            "next",
+            format!(
+                "{}/api/v1/timelines/public?limit={}&max_id={}",
+                base_url,
+                limit,
+                get_key(collection.last().unwrap())
+            ),
+        );
+        let prev = (
+            "prev",
+            format!(
+                "{}/api/v1/timelines/public?limit={}&since_id={}",
+                base_url,
+                limit,
+                get_key(collection.first().unwrap())
+            ),
+        );
+        if collection.len() >= limit && limit > 0 {
+            Some(Link(vec![next, prev]))
+        } else {
+            Some(Link(vec![prev]))
+        }
     }
 }
